requirements.txt
streamlit
tensorflow
pandas
Pillow
gdown